method2_production/requirements.txt

fastapi
uvicorn 
supabase
groq
python-dotenv
pydantic